+++ DataAccessLayer/Abstract/ICategoryRepository.cs
@@ -5,8 +5,8 @@ namespace DataAccessLayer.Abstract
     public interface ICategoryRepository : IRepository<Category>    
     {
         List<Product> GetPopularCategories();
-        Category GetByIdWithProducts(int categoryId);
-        void DeleteFromCategory(int productId, int categoryId);
+        Category GetByIdWithProducts(double categoryId);
+        void DeleteFromCategory(double productId, double categoryId);
         bool GetByName(string name);
     }
 }
