+++ DataAccessLayer/Abstract/IRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IRepository<T>
     {
-        T GetById(int id);
+        T GetById(double id);
         List<T> GetAll();
         void Create(T entity);
         void Update(T entity);
