+++ DataAccessLayer/Concrete/EFCore/EFCoreCategoryRepository.cs
@@ -15,13 +15,13 @@ namespace DataAccessLayer.Concrete.EFCore
         {
             get { return _context as ShopContext; }         
         }
-        public void DeleteFromCategory(int productId, int categoryId)
+        public void DeleteFromCategory(double productId, double categoryId)
         {
             var cmd = "delete from productcategory where ProductId=@p0 and CategoryId=@p1";
             ShopContext.Database.ExecuteSqlRaw(cmd, productId, categoryId);
         }
 
-        public Category GetByIdWithProducts(int categoryId)
+        public Category GetByIdWithProducts(double categoryId)
         {
             return ShopContext.Categories
                 .Where(p => p.CategoryId == categoryId)
