+++ PresentationLayer/Models/Cart/CartItemModel.cs
@@ -2,9 +2,9 @@
 {
     public class CartItemModel
     {
-        public int CartItemId { get; set; }
+        public double CartItemId { get; set; }
         public int ProductQuantity { get; set; }
-        public int ProductId { get; set; }
+        public double ProductId { get; set; }
         public string? Name { get; set; }
         public double Price { get; set; }
         public string? ImageUrl { get; set; }
