+++ PresentationLayer/Models/Cart/CartModel.cs
@@ -5,7 +5,7 @@ namespace PresentationLayer.Models
 {
     public class CartModel
     {
-        public int CartId { get; set; }
+        public double CartId { get; set; }
         public ProductModel ProductModel { get; set; }
         public List<CartItemModel>? CartItems { get; set; }
         public decimal TotalPrice()
