+++ PresentationLayer/Models/Order/OrderItemModel.cs
@@ -4,7 +4,7 @@ namespace PresentationLayer.Models
 {
     public class OrderItemModel
     {
-        public int OrderItemId { get; set; }
+        public double OrderItemId { get; set; }
         public double Price { get; set; }
         public string Name { get; set; }
         public string ImageUrl { get; set; }
